refactor: 서비스에서 사용하는 dto는 서비스 계층으로 이동
