+++ src/main/java/com/airdnb/chat/ChatRoomController.java
@@ -1,9 +1,9 @@
 package com.airdnb.chat;
 
-import com.airdnb.chat.dto.ChatRoomCreation;
 import com.airdnb.chat.dto.ChatRoomCreationRequest;
 import com.airdnb.chat.dto.ChatRoomResponse;
 import com.airdnb.chat.dto.MessageResponse;
+import com.airdnb.chat.service.ChatRoomCreation;
 import com.airdnb.chat.service.ChatRoomService;
 import com.airdnb.global.ApiResponse;
 import com.airdnb.global.UriMaker;
